+++ test/Adapter/Factory/AzureAdapterFactoryTest.php
@@ -30,6 +30,8 @@ class AzureAdapterFactoryTest extends TestCase
 
     public function testCreateService()
     {
+        $this->markTestSkipped("Skipped due to https://github.com/thephpleague/flysystem-azure/pull/16");
+        
         $sm = Bootstrap::getServiceManager();
         $manager = $sm->get('BsbFlysystemAdapterManager');